+++ src/hooks/useWalletWidgets.tsx
@@ -3,16 +3,22 @@ import { useCallback, useMemo } from "react";
 import { createExternalWallet } from "@/core";
 import { ExternalConnector } from "@/core/types";
 
-export function useWalletWidgets(connectors: any, config: any) {
+export function useWalletWidgets(connectors: any, config: any, onError?: (e: Error) => void) {
   const createChainWidget = useCallback(
     (chainId: string, externalConnectors: ExternalConnector[]) => (
       <>
         {externalConnectors.map(({ id, widget: Component }) => (
-          <Component key={id} id={id} connector={connectors[chainId]} createWallet={createExternalWallet} />
+          <Component
+            key={id}
+            id={id}
+            connector={connectors[chainId]}
+            createWallet={createExternalWallet}
+            onError={onError}
+          />
         ))}
       </>
     ),
-    [connectors],
+    [connectors, onError],
   );
 
   return useMemo(